// frontend/types/api/contacts/models/EmailContactUpdate.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Update model for email contacts.
 */
export type EmailContactUpdate = {
    display_name?: (string | null);
    given_name?: (string | null);
    family_name?: (string | null);
    tags?: (Array<string> | null);
    notes?: (string | null);
};